implementirano dohvatanje sigurnosnog pitanja iz baze
